Fixed some grammar and sizing issues
Changed the wording of some of the descriptions and fixed some grammar mistakes on multiple pages. There were also some sizing issues which would block use of some buttons on the index page which are hopefully fixed. Also, there were some sizing differences on the chain game page which now align with the sizes on other pages.